Properly order Sprockets asset paths.
Prior to this change we were always inserting `app/` and `config/` to
the beginning of the `Sprockets::Railtie.config.assets.paths` array.

Unfortunately, this has the negative impact of matching the lookup of
`app/assets/images` and `app/assets/stylesheets` (any other folders
under `app/` actually) with a root of `app/` (instead of a root of
`app/assets`). This caused the issue detailed in #96 (stylesheets being
available at `/assets/assets/application.css` instead of
`/assets/application.css`).

This change forces `app/` and `config/` to be inserted after any folders
under `app/` other than `app/assets/javascripts` (which we remove in
`ember:bootstrap`, but is still dealt with here).

Closes #96.